Move client state back to _clientwindow
Refactoring showed that client state should indeed be handled by the
client rather than the connection to the lobby server.
