+++ tests/unit_tests/client/test_updating.py
@@ -6,13 +6,13 @@ import config
 def test_client_sends_current_version(qtbot, mocker):
     import client
     c = client.instance
-    mocker.patch.object(c.lobby_server, 'send')
-    mocker.patch.object(c.lobby_server, 'connected')
-    mocker.patch.object(c.lobby_server, 'socket')
+    mocker.patch.object(c.lobby_connection, 'send')
+    mocker.patch.object(c.lobby_connection, 'connected')
+    mocker.patch.object(c.lobby_connection, 'socket')
 
-    c.lobby_server.on_connected()
+    c.on_connected()
 
-    args, kwargs = c.lobby_server.send.call_args
+    args, kwargs = c.lobby_connection.send.call_args
     assert args[0]['version'] == config.VERSION
